bug fix: was not including all mime types / extensions
Ran into issue of not serving a specific mimetype, based on
the file extension. For instance, the code below was used to
assign a mimetype to js files:
{
...
	javascript: 'text/javascript',
...
}

This led to js files needing the extension '.javascript' to work.

Changed this to '.js' and have included a could other image types.